Rename test package to mocks according to standard package layout
