+++ auth/middleware_test.go
@@ -7,7 +7,7 @@ package auth
 import (
 	"context"
 	"github.com/lestrrat-go/jwx/jwa"
-	"github.com/sap/cloud-security-client-go/test"
+	"github.com/sap/cloud-security-client-go/mocks"
 	"io/ioutil"
 	"net/http"
 	"net/http/httptest"
@@ -25,7 +25,7 @@ func TestEnd2End(t *testing.T) {
 	tests := []struct {
 		name    string
 		header  map[string]interface{}
-		claims  test.OIDCClaims
+		claims  mocks.OIDCClaims
 		wantErr bool
 	}{
 		{
@@ -36,7 +36,7 @@ func TestEnd2End(t *testing.T) {
 		}, {
 			name:   "valid with aud array",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Audience("notMyClient", oidcMockServer.Config.ClientID).
 				Build(),
 			wantErr: false,
@@ -44,14 +44,14 @@ func TestEnd2End(t *testing.T) {
 		{
 			name:   "valid with single aud",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Audience(oidcMockServer.Config.ClientID).
 				Build(),
 			wantErr: false,
 		},
 		{
 			name: "no key id in token",
-			header: test.NewOIDCHeaderBuilder(oidcMockServer.DefaultHeaders()).
+			header: mocks.NewOIDCHeaderBuilder(oidcMockServer.DefaultHeaders()).
 				KeyID("").
 				Build(),
 			claims:  oidcMockServer.DefaultClaims(),
@@ -59,132 +59,132 @@ func TestEnd2End(t *testing.T) {
 		}, {
 			name:   "expired",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				ExpiresAt(time.Now().Add(-1 * time.Minute)).
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "no expiry provided",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				WithoutExpiresAt().
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "before validity",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				NotBefore(time.Now().Add(1 * time.Minute)).
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "wrong audience",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Audience("notMyClient").
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "wrong audience array",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Audience("notMyClient", "neitherThisOne").
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "wrong issuer",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Issuer("https://another.oidc-server.com/").
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "no http/s prefix for issuer",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Issuer("127.0.0.1:64004").
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "issuer malicious",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Issuer(oidcMockServer.Server.URL + "?redirect=https://malicious.ondemand.com/tokens%3Ftenant=9451dd2etrial").
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "issuer malicious2",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Issuer(oidcMockServer.Server.URL + "\\\\@malicious.ondemand.com").
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "issuer malicious3",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Issuer(oidcMockServer.Server.URL + "@malicious.ondemand.com").
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "issuer malicious4",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Issuer("https://malicious.ondemand.com/token_keys///" + oidcMockServer.Server.URL).
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "issuer malicious5",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Issuer("https://malicious.ondemand.com/token_keys@" + strings.TrimPrefix(oidcMockServer.Server.URL, "https://")).
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "issuer malicious6",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Issuer(oidcMockServer.Server.URL + "///malicious.ondemand.com/token_keys").
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "issuer malicious7",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Issuer(oidcMockServer.Server.URL + "\\\\@malicious.ondemand.com/token_keys").
 				Build(),
 			wantErr: true,
 		}, {
 			name:   "issuer empty",
 			header: oidcMockServer.DefaultHeaders(),
-			claims: test.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
+			claims: mocks.NewOIDCClaimsBuilder(oidcMockServer.DefaultClaims()).
 				Issuer("").
 				Build(),
 			wantErr: true,
 		}, {
 			name: "wrong key id",
-			header: test.NewOIDCHeaderBuilder(oidcMockServer.DefaultHeaders()).
+			header: mocks.NewOIDCHeaderBuilder(oidcMockServer.DefaultHeaders()).
 				KeyID("wrongKey").
 				Build(),
 			claims:  oidcMockServer.DefaultClaims(),
 			wantErr: true,
 		}, {
 			name: "none algorithm",
-			header: test.NewOIDCHeaderBuilder(oidcMockServer.DefaultHeaders()).
+			header: mocks.NewOIDCHeaderBuilder(oidcMockServer.DefaultHeaders()).
 				Alg(jwa.NoSignature).
 				Build(),
 			claims:  oidcMockServer.DefaultClaims(),
 			wantErr: true,
 		}, {
 			name: "empty algorithm",
-			header: test.NewOIDCHeaderBuilder(oidcMockServer.DefaultHeaders()).
+			header: mocks.NewOIDCHeaderBuilder(oidcMockServer.DefaultHeaders()).
 				Alg("").
 				Build(),
 			claims:  oidcMockServer.DefaultClaims(),
 			wantErr: true,
 		}, {
 			name: "wrong algorithm",
-			header: test.NewOIDCHeaderBuilder(oidcMockServer.DefaultHeaders()).
+			header: mocks.NewOIDCHeaderBuilder(oidcMockServer.DefaultHeaders()).
 				Alg(jwa.HS256).
 				Build(),
 			claims:  oidcMockServer.DefaultClaims(),
@@ -231,8 +231,8 @@ func GetTestHandler() http.HandlerFunc {
 	}
 }
 
-func GetTestServer() (clientServer *httptest.Server, oidcServer *test.MockServer) {
-	mockServer, _ := test.NewOIDCMockServer()
+func GetTestServer() (clientServer *httptest.Server, oidcServer *mocks.MockServer) {
+	mockServer, _ := mocks.NewOIDCMockServer()
 	options := Options{
 		ErrorHandler: nil,
 		HTTPClient:   mockServer.Server.Client(),
