+++ go.mod
@@ -3,11 +3,11 @@ module github.com/sap/cloud-security-client-go
 go 1.15
 
 require (
-	github.com/google/uuid v1.1.2
+	github.com/google/uuid v1.2.0
 	github.com/gorilla/handlers v1.5.1
 	github.com/gorilla/mux v1.8.0
 	github.com/lestrrat-go/jwx v1.1.7
 	github.com/patrickmn/go-cache v2.1.0+incompatible
-	github.com/pquerna/cachecontrol v0.0.0-20200921180117-858c6e7e6b7e
-	golang.org/x/sync v0.0.0-20201020160332-67f06af15bc9
+	github.com/pquerna/cachecontrol v0.1.0
+	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
 )
